+++ R/varImp.R
@@ -8,11 +8,14 @@
 #' than one permutation (default is 10) the average of the decrease in training
 #' AUC is computed.
 #'
-#' @param model \linkS4class{SDMmodel} object.
+#' @param model \linkS4class{SDMmodel} or \linkS4class{SDMmodelCV} object.
 #' @param permut integer. Number of permutations, default is 10.
 #'
-#' @details Note that it could return values slightly different from MaxEnt java
+#' @details Note that it could return values slightly different from MaxEnt Java
 #' software due to a different random permutation.
+#'
+#' For \code{\link{SDMmodelCV}} objects the function returns the average and the
+#' standard deviation of the permutation importances of the single models.
 #'
 #' @return data.frame with the ordered permutation importance.
 #' @export
@@ -54,23 +57,52 @@
 #' }
 varImp <- function(model, permut = 10) {
 
-  set.seed(25)
   vars <- colnames(model@p@data)
-  model_auc <- auc(model)
-  permuted_auc <- matrix(nrow = permut, ncol = length(vars))
 
-  n_pres <- nrow(model@p@data)
+  if (class(model) == "SDMmodel") {
+    total <- length(vars)
+  } else {
+    l <- length(model@models)
+    total <- length(vars) * l
+  }
 
   pb <- progress::progress_bar$new(
     format = "Variable importance [:bar] :percent in :elapsedfull",
-    total = length(vars),
-    clear = FALSE, width = 60, show_after = 0)
+    total = total, clear = FALSE, width = 60, show_after = 0)
   pb$tick(0)
+
+  if (class(model) == "SDMmodel") {
+    model_auc <- auc(model)
+    output <- .compute_permutation(model, model_auc, vars, permut, pb)
+  } else {
+    pis <- matrix(nrow = length(vars), ncol = l)
+    for (i in 1:l) {
+      model_auc <- auc(model@models[[i]])
+      df <- .compute_permutation(model@models[[i]], model_auc, vars, permut, pb)
+      index <- match(df[, 1], vars)
+      pis[, i] <- df[order(index), 2]
+    }
+    output <- data.frame(Variable = vars,
+                         Permutation_importance = rowMeans(pis),
+                         sd = round(apply(pis, 1, sd), 3),
+                         stringsAsFactors = FALSE)
+  }
+
+  output <- output[order(output$Permutation_importance, decreasing = TRUE), ]
+  row.names(output) <- NULL
+
+  return(output)
+}
+
+.compute_permutation <- function(model, model_auc, vars, permut, pb) {
+
+  permuted_auc <- matrix(nrow = permut, ncol = length(vars))
+  n_pres <- nrow(model@p@data)
+  set.seed(25)
 
   for (j in 1:length(vars)) {
     for (i in 1:permut) {
-      data <- sample(c(model@p@data[, vars[j]],
-                       model@a@data[, vars[j]]))
+      data <- sample(c(model@p@data[, vars[j]], model@a@data[, vars[j]]))
       if (is.factor(model@p@data[, vars[j]]))
         data <- as.factor(data)
       p_copy <- model@p
@@ -95,8 +127,6 @@ varImp <- function(model, permut = 10) {
                        stringsAsFactors = FALSE)
   if (permut > 1)
     output$sd <- round(sd_auc, 3)
-  output <- output[order(output$Permutation_importance, decreasing = TRUE), ]
-  row.names(output) <- NULL
 
   return(output)
 }